refactor(ai): split planning_helpers into topic libraries (#3941)
Move the near-gate planning helpers monolithic into focused modules with a
thin barrel re-export so the 1000-line gate has headroom, remirror tests to
topic files, and point ai_dedup allowlists at the new canonical homes.
