+++ packages/colonizethis_ai/test/planning/planning_helpers_test.dart
@@ -1,1034 +1,15 @@
-// Unit tests for `planning_helpers.dart` (Refs #3278).
-//
-// Pins the shared dedup helpers:
-//   - `gpFactionIdsAtWarWith` — GP-only filter, sort/determinism, non-GP exclusion
-//   - `isAtWarWithAnyGreatPower` — GP-war presence check, non-GP exclusion,
-//     empty-set false, equivalence with `gpFactionIdsAtWarWith.isNotEmpty`
-//     (Refs #3717)
-//   - `isOwnOldWorldExpansionStalled` / `isOwnOldWorldBelowConquestQuota` —
-//     own-OW snapshot projections: stall-band / quota boundaries, negative
-//     cases, equivalence with the underlying int predicates (Refs #3717)
-//   - `oldWorldProvinceLeadOver` — another faction's OW province count minus the
-//     active player's own (`provinceCountOwnedBy - oldWorldProvincesOwned`):
-//     positive lead / negative deficit / zero, owner isolation, unowned-faction
-//     pure deficit, determinism (Refs #3717)
-//   - `gpAtWarPeaceTargetsWhere` — predicate-filtered GP at-war peace-target
-//     collector: keep-subset filtering, ascending sort/determinism, non-GP
-//     exclusion, keep-all/keep-none edges, one keep call per GP (Refs #3717)
-//   - `minorAtWarPeaceTargetsWhere` — predicate-filtered minor at-war
-//     peace-target collector: null-keep keep-all, keep-subset filtering,
-//     GP/tribe exclusion, keep-none edge, ascending sort/determinism
-//     (Refs #3717)
-//   - `tribeAtWarPeaceTargetsWhere` — predicate-filtered tribe at-war
-//     peace-target collector: null-keep keep-all, keep-subset filtering,
-//     GP/minor exclusion, keep-none edge, ascending sort/determinism
-//     (Refs #3717)
-//   - `peaceTargetsExcludingBlocker` — drop-blocker-then-sort peace-target
-//     skeleton: blocker exclusion, null/absent blocker keep-all, empty input,
-//     input-order independence, Set (atWarWith) input (Refs #3717)
-//   - `scaleWeightedBonus` — `<= 0.0` guard, `> 1.0` clamp, rounding
-//   - `resolvePhaseColonialPressureActive` — COLONIAL-only gate
-//   - `resolvePhaseExpandOrColonialLiteActive` — EXPAND / COLONIAL-lite gate
-//
-// Invadable-province ownership helpers (`anyInvadableProvinceOwnedByMinor`,
-// `anyInvadableProvinceOwnedByGreatPower`, `factionOwnsInvadableOldWorldProvince`,
-// `addInvadableProvinceMinorOwnersNotAtWar`) are pinned in the sibling
-// `planning_helpers_part2_test.dart`; the diplomatic-cooldown and at-war
-// peace/order scoring helpers (`hasRecentDiplomaticEventWithinCooldown`,
-// `atWarPeaceTargetBonus`, `atWarGreatPowerOrderTarget`) are pinned in
-// `planning_helpers_part3_test.dart`, to keep each file within the repo
-// non-comment line limit.
+// Smoke coverage for the `planning_helpers.dart` barrel (Refs #3941).
+// Topic unit contracts live in `planning_*_test.dart` siblings; this file only
+// pins that the barrel re-exports remain resolvable after the topic split.
 
-import 'package:colonizethis_ai/src/perception/perception_snapshot.dart';
-import 'package:colonizethis_ai/src/planning/observer_goal_phase.dart';
-import 'package:colonizethis_ai/src/planning/phase_planner_dispatch.dart'
-    show PhasePlanOutcome;
-import 'package:colonizethis_ai/src/planning/phase_priority_weights.dart'
-    show PhasePriorityWeights;
 import 'package:colonizethis_ai/src/planning/planning_helpers.dart';
-import 'package:colonizethis_data/colonizethis_data.dart'
-    show
-        isBelowObserverConquestQuota,
-        isStalledOldWorldExpansion,
-        kObserverConquestMinOwProvincesPerGp,
-        kStalledOldWorldProvinceThreshold;
-import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
-// Distinct per-field values so a slot mix-up (e.g. a projection that
-// returns `oldWorldCivilian` instead of `oldWorldConquest`) fails hard
-// rather than silently matching a shared default.
-const PhasePriorityWeights _weights = PhasePriorityWeights(
-  oldWorldConquest: 0.11,
-  newWorldAcquisition: 0.22,
-  oldWorldCivilian: 0.33,
-  newWorldCivilian: 0.44,
-);
-
-const String _gp1 = 'gp1';
-const String _gp2 = 'gp2';
-const String _gp3 = 'gp3';
-const String _tribe1 = 'tribe1';
-const String _minor1 = 'minor1';
-
-Game _gameWithGps() {
-  return Game(
-    id: 'g-3278-planning-helpers',
-    worldState: WorldState(
-      turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-      oldWorld: const RegionData(provinces: []),
-      newWorld: const RegionData(provinces: []),
-    ),
-    players: const [
-      Player(id: _gp1, displayName: 'GP1', isHuman: false),
-      Player(id: _gp2, displayName: 'GP2', isHuman: false),
-      Player(id: _gp3, displayName: 'GP3', isHuman: false),
-    ],
-    minorNations: const [MinorNation(id: _minor1, displayName: 'Minor1')],
-    tribes: const [Tribe(id: _tribe1, displayName: 'Tribe1')],
-  );
-}
-
-AIWorldSnapshot _snapshotWithAtWar(List<String> atWarWith) {
-  return AIWorldSnapshot(
-    playerId: _gp1,
-    threats: ThreatSummary(atWarWith: atWarWith),
-    opportunities: const OpportunitySummary(),
-    conquest: const ConquestSummary(),
-    economy: const EconomySummary(),
-    relations: const {},
-  );
-}
-
-AIWorldSnapshot _snapshotWithOwnOw(int oldWorldProvincesOwned) {
-  return AIWorldSnapshot(
-    playerId: _gp1,
-    threats: const ThreatSummary(atWarWith: []),
-    opportunities: const OpportunitySummary(),
-    conquest: ConquestSummary(oldWorldProvincesOwned: oldWorldProvincesOwned),
-    economy: const EconomySummary(),
-    relations: const {},
-  );
-}
-
-// Game where [ownerCounts] maps a factionId to the number of Old World
-// provinces it owns, so `provinceCountOwnedBy` (and hence the
-// `oldWorldProvinceLeadOver` projection under test) returns those counts
-// deterministically via the memoised province-owner cache.
-Game _gameOwning(Map<String, int> ownerCounts) {
-  final provinces = <Province>[
-    for (final entry in ownerCounts.entries)
-      for (var i = 0; i < entry.value; i++)
-        Province(
-          id: 'oldWorld|${entry.key}_$i',
-          regionId: 'oldWorld',
-          ownerId: entry.key,
-        ),
-  ];
-  return Game(
-    id: 'g-3717-ow-lead',
-    worldState: WorldState(
-      turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-      oldWorld: RegionData(provinces: provinces),
-      newWorld: const RegionData(provinces: []),
-    ),
-    players: [
-      for (final id in ownerCounts.keys)
-        Player(id: id, displayName: id.toUpperCase(), isHuman: false),
-    ],
-  );
-}
-
-// Game whose Great Powers each have [techCountByPlayer] unlocked techs (the
-// `techUnlocked` map carries that many `true` flags), so `unlockedTechCount`
-// and the `isPursuingTechStealPosture` deficit comparison under test resolve
-// deterministically. Minor nations and tribes carry no tech state.
-Game _gameWithTechs(Map<String, int> techCountByPlayer) {
-  Map<String, bool> techs(int count) => {
-    for (var i = 0; i < count; i++) 'tech_$i': true,
-  };
-  return Game(
-    id: 'g-3793-tech-steal',
-    worldState: WorldState(
-      turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-      oldWorld: const RegionData(provinces: []),
-      newWorld: const RegionData(provinces: []),
-    ),
-    players: [
-      for (final entry in techCountByPlayer.entries)
-        Player(
-          id: entry.key,
-          displayName: entry.key.toUpperCase(),
-          isHuman: false,
-          techUnlocked: techs(entry.value),
-        ),
-    ],
-    minorNations: const [MinorNation(id: _minor1, displayName: 'Minor1')],
-    tribes: const [Tribe(id: _tribe1, displayName: 'Tribe1')],
-  );
-}
-
 void main() {
-  group('unlockedTechCount (Refs #3793)', () {
-    test('counts only true-flagged techs; null/empty map is 0', () {
-      expect(
-        unlockedTechCount(
-          const Player(id: _gp1, displayName: 'GP1', isHuman: false),
-        ),
-        0,
-      );
-      expect(
-        unlockedTechCount(
-          const Player(
-            id: _gp1,
-            displayName: 'GP1',
-            isHuman: false,
-            techUnlocked: {'a': true, 'b': true, 'c': false},
-          ),
-        ),
-        2,
-      );
-    });
+  test('barrel re-exports topic helpers', () {
+    expect(clampPhaseWeightUpperUnit(2.0), 1.0);
+    expect(planningListEquals(const ['a'], const ['a']), isTrue);
+    expect(resolvePhaseColonialPressureActive, isNotNull);
+    expect(gpFactionIdsAtWarWith, isNotNull);
   });
-
-  group('isPursuingTechStealPosture (Refs #3793, AC4c)', () {
-    test('true when a rival GP leads by at least the deficit', () {
-      // gp1 owns 2 techs; gp2 owns 4 → lead 2 >= deficit 1 → posture.
-      final game = _gameWithTechs({_gp1: 2, _gp2: 4, _gp3: 1});
-      expect(isPursuingTechStealPosture(game, _gp1), isTrue);
-    });
-
-    test('true at exactly the default deficit (lead of 1)', () {
-      final game = _gameWithTechs({_gp1: 3, _gp2: 4});
-      expect(isPursuingTechStealPosture(game, _gp1), isTrue);
-    });
-
-    test('false when the active GP leads or is tied with every rival', () {
-      // gp1 is the most advanced; no rival leads it.
-      final leadGame = _gameWithTechs({_gp1: 5, _gp2: 4, _gp3: 2});
-      expect(isPursuingTechStealPosture(leadGame, _gp1), isFalse);
-      // Parity with the top rival → lead 0 < deficit 1 → no posture.
-      final tiedGame = _gameWithTechs({_gp1: 4, _gp2: 4});
-      expect(isPursuingTechStealPosture(tiedGame, _gp1), isFalse);
-    });
-
-    test('false when there are no rival Great Powers', () {
-      final game = _gameWithTechs({_gp1: 0});
-      expect(isPursuingTechStealPosture(game, _gp1), isFalse);
-    });
-
-    test('false for an unknown active player id', () {
-      final game = _gameWithTechs({_gp1: 1, _gp2: 9});
-      expect(isPursuingTechStealPosture(game, 'nobody'), isFalse);
-    });
-
-    test('ignores minor nations and tribes (not Player rivals)', () {
-      // Only gp1 is a Player; minors/tribes carry no tech state and are not
-      // rivals, so a lone GP is never behind.
-      final game = _gameWithTechs({_gp1: 0});
-      expect(isPursuingTechStealPosture(game, _gp1), isFalse);
-    });
-
-    test('deterministic for fixed inputs', () {
-      final game = _gameWithTechs({_gp1: 1, _gp2: 5});
-      final a = isPursuingTechStealPosture(game, _gp1);
-      final b = isPursuingTechStealPosture(game, _gp1);
-      expect(a, isTrue);
-      expect(b, a);
-    });
-  });
-
-  group('gpFactionIdsAtWarWith', () {
-    test('filters to Great Powers only and sorts ascending', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_gp3, _tribe1, _gp1, _minor1, _gp2]);
-      expect(gpFactionIdsAtWarWith(game, snapshot), [_gp1, _gp2, _gp3]);
-    });
-
-    test('returns empty when no GP wars are active', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_tribe1, _minor1]);
-      expect(gpFactionIdsAtWarWith(game, snapshot), isEmpty);
-    });
-
-    test('sorts regardless of atWarWith iteration order', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_gp3, _gp2, _gp1]);
-      final a = gpFactionIdsAtWarWith(game, snapshot);
-      final b = gpFactionIdsAtWarWith(game, snapshot);
-      expect(a, [_gp1, _gp2, _gp3]);
-      expect(b, a);
-    });
-  });
-
-  group('isAtWarWithAnyGreatPower (Refs #3717)', () {
-    test('true when at least one at-war faction is a Great Power', () {
-      final game = _gameWithGps();
-      expect(
-        isAtWarWithAnyGreatPower(game, _snapshotWithAtWar([_tribe1, _gp2])),
-        isTrue,
-      );
-    });
-
-    test('false when no at-war faction resolves to a Great Power', () {
-      final game = _gameWithGps();
-      expect(
-        isAtWarWithAnyGreatPower(game, _snapshotWithAtWar([_tribe1, _minor1])),
-        isFalse,
-      );
-    });
-
-    test('false on an empty atWarWith set', () {
-      final game = _gameWithGps();
-      expect(
-        isAtWarWithAnyGreatPower(game, _snapshotWithAtWar(const [])),
-        isFalse,
-      );
-    });
-
-    test('agrees with gpFactionIdsAtWarWith.isNotEmpty (equivalence)', () {
-      final game = _gameWithGps();
-      for (final atWar in <List<String>>[
-        const [],
-        [_tribe1],
-        [_minor1, _tribe1],
-        [_gp1],
-        [_gp3, _tribe1, _gp1, _minor1, _gp2],
-      ]) {
-        final snapshot = _snapshotWithAtWar(atWar);
-        expect(
-          isAtWarWithAnyGreatPower(game, snapshot),
-          gpFactionIdsAtWarWith(game, snapshot).isNotEmpty,
-          reason: 'mismatch for atWarWith=$atWar',
-        );
-      }
-    });
-  });
-
-  group('isOwnOldWorldExpansionStalled (Refs #3717)', () {
-    test('true inside the stalled OW band (0 < ow <= threshold)', () {
-      expect(isOwnOldWorldExpansionStalled(_snapshotWithOwnOw(1)), isTrue);
-      expect(
-        isOwnOldWorldExpansionStalled(
-          _snapshotWithOwnOw(kStalledOldWorldProvinceThreshold),
-        ),
-        isTrue,
-      );
-    });
-
-    test('false at zero OW and above the stall band (negative cases)', () {
-      expect(isOwnOldWorldExpansionStalled(_snapshotWithOwnOw(0)), isFalse);
-      expect(
-        isOwnOldWorldExpansionStalled(
-          _snapshotWithOwnOw(kStalledOldWorldProvinceThreshold + 1),
-        ),
-        isFalse,
-      );
-    });
-
-    test('matches isStalledOldWorldExpansion(ownOw) across a range', () {
-      for (var ow = 0; ow <= kStalledOldWorldProvinceThreshold + 3; ow++) {
-        expect(
-          isOwnOldWorldExpansionStalled(_snapshotWithOwnOw(ow)),
-          isStalledOldWorldExpansion(ow),
-          reason: 'mismatch for ownOw=$ow',
-        );
-      }
-    });
-  });
-
-  group('isOwnOldWorldBelowConquestQuota (Refs #3717)', () {
-    test('true below the observer conquest quota', () {
-      expect(isOwnOldWorldBelowConquestQuota(_snapshotWithOwnOw(0)), isTrue);
-      expect(
-        isOwnOldWorldBelowConquestQuota(
-          _snapshotWithOwnOw(kObserverConquestMinOwProvincesPerGp - 1),
-        ),
-        isTrue,
-      );
-    });
-
-    test('false at and above the quota (boundary negative cases)', () {
-      expect(
-        isOwnOldWorldBelowConquestQuota(
-          _snapshotWithOwnOw(kObserverConquestMinOwProvincesPerGp),
-        ),
-        isFalse,
-      );
-      expect(
-        isOwnOldWorldBelowConquestQuota(
-          _snapshotWithOwnOw(kObserverConquestMinOwProvincesPerGp + 5),
-        ),
-        isFalse,
-      );
-    });
-
-    test('matches isBelowObserverConquestQuota(ownOw) across a range', () {
-      for (var ow = 0; ow <= kObserverConquestMinOwProvincesPerGp + 3; ow++) {
-        expect(
-          isOwnOldWorldBelowConquestQuota(_snapshotWithOwnOw(ow)),
-          isBelowObserverConquestQuota(ow),
-          reason: 'mismatch for ownOw=$ow',
-        );
-      }
-    });
-  });
-
-  group('oldWorldProvinceLeadOver (Refs #3717)', () {
-    test('positive when the faction owns more than the own OW count', () {
-      final game = _gameOwning({_gp2: 5});
-      expect(
-        oldWorldProvinceLeadOver(
-          game: game,
-          snapshot: _snapshotWithOwnOw(3),
-          factionId: _gp2,
-        ),
-        2,
-      );
-    });
-
-    test('negative (deficit view) when the faction owns fewer', () {
-      final game = _gameOwning({_gp2: 1});
-      expect(
-        oldWorldProvinceLeadOver(
-          game: game,
-          snapshot: _snapshotWithOwnOw(4),
-          factionId: _gp2,
-        ),
-        -3,
-      );
-    });
-
-    test('zero when the faction owns exactly the own OW count', () {
-      final game = _gameOwning({_gp2: 4});
-      expect(
-        oldWorldProvinceLeadOver(
-          game: game,
-          snapshot: _snapshotWithOwnOw(4),
-          factionId: _gp2,
-        ),
-        0,
-      );
-    });
-
-    test('counts only the queried faction (other owners excluded)', () {
-      final game = _gameOwning({_gp2: 6, _gp3: 9});
-      expect(
-        oldWorldProvinceLeadOver(
-          game: game,
-          snapshot: _snapshotWithOwnOw(2),
-          factionId: _gp2,
-        ),
-        4,
-      );
-    });
-
-    test('unowned faction yields a pure deficit (count 0)', () {
-      final game = _gameOwning({_gp2: 3});
-      expect(
-        oldWorldProvinceLeadOver(
-          game: game,
-          snapshot: _snapshotWithOwnOw(5),
-          factionId: _gp3,
-        ),
-        -5,
-      );
-    });
-
-    test('deterministic for fixed inputs', () {
-      final game = _gameOwning({_gp2: 7});
-      final snapshot = _snapshotWithOwnOw(3);
-      final a = oldWorldProvinceLeadOver(
-        game: game,
-        snapshot: snapshot,
-        factionId: _gp2,
-      );
-      final b = oldWorldProvinceLeadOver(
-        game: game,
-        snapshot: snapshot,
-        factionId: _gp2,
-      );
-      expect(a, 4);
-      expect(b, a);
-    });
-  });
-
-  group('gpAtWarPeaceTargetsWhere (Refs #3717)', () {
-    test('keeps only GP at-war factions matching the predicate, sorted', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_gp3, _tribe1, _gp1, _minor1, _gp2]);
-      expect(
-        gpAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (factionId) => factionId != _gp2,
-        ),
-        [_gp1, _gp3],
-      );
-    });
-
-    test('keep-all equals gpFactionIdsAtWarWith (GP filter + sort)', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_gp3, _tribe1, _gp1, _minor1, _gp2]);
-      expect(
-        gpAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => true,
-        ),
-        gpFactionIdsAtWarWith(game, snapshot),
-      );
-    });
-
-    test('keep-none returns empty', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_gp1, _gp2, _gp3]);
-      expect(
-        gpAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => false,
-        ),
-        isEmpty,
-      );
-    });
-
-    test('never offers a non-GP even when the predicate would keep it', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_tribe1, _minor1]);
-      expect(
-        gpAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => true,
-        ),
-        isEmpty,
-      );
-    });
-
-    test('result is sorted ascending regardless of atWarWith order', () {
-      final game = _gameWithGps();
-      expect(
-        gpAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: _snapshotWithAtWar([_gp3, _gp1, _gp2]),
-          keep: (_) => true,
-        ),
-        [_gp1, _gp2, _gp3],
-      );
-    });
-
-    test('invokes keep exactly once per at-war GP in ascending order', () {
-      final game = _gameWithGps();
-      final snapshot = _snapshotWithAtWar([_gp3, _tribe1, _gp1, _gp2]);
-      final seen = <String>[];
-      gpAtWarPeaceTargetsWhere(
-        game: game,
-        snapshot: snapshot,
-        keep: (factionId) {
-          seen.add(factionId);
-          return true;
-        },
-      );
-      expect(seen, [_gp1, _gp2, _gp3]);
-    });
-  });
-
-  group('minorAtWarPeaceTargetsWhere (Refs #3717)', () {
-    Game gameWithMinors() => Game(
-      id: 'g-3717-minor-peace',
-      worldState: WorldState(
-        turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-        oldWorld: const RegionData(provinces: []),
-        newWorld: const RegionData(provinces: []),
-      ),
-      players: const [Player(id: _gp1, displayName: 'GP1', isHuman: false)],
-      minorNations: const [
-        MinorNation(id: 'minorA', displayName: 'MinorA'),
-        MinorNation(id: 'minorB', displayName: 'MinorB'),
-        MinorNation(id: 'minorC', displayName: 'MinorC'),
-      ],
-      tribes: const [Tribe(id: _tribe1, displayName: 'Tribe1')],
-    );
-
-    test('keep == null keeps every at-war minor, sorted ascending', () {
-      final game = gameWithMinors();
-      final snapshot = _snapshotWithAtWar([
-        'minorC',
-        _gp1,
-        'minorA',
-        _tribe1,
-        'minorB',
-      ]);
-      expect(minorAtWarPeaceTargetsWhere(game: game, snapshot: snapshot), [
-        'minorA',
-        'minorB',
-        'minorC',
-      ]);
-    });
-
-    test('keeps only minors matching the predicate, sorted', () {
-      final game = gameWithMinors();
-      final snapshot = _snapshotWithAtWar(['minorC', 'minorA', 'minorB']);
-      expect(
-        minorAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (factionId) => factionId != 'minorB',
-        ),
-        ['minorA', 'minorC'],
-      );
-    });
-
-    test('never offers a GP or tribe even with a keep-all predicate', () {
-      final game = gameWithMinors();
-      final snapshot = _snapshotWithAtWar([_gp1, _tribe1, 'minorA']);
-      expect(
-        minorAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => true,
-        ),
-        ['minorA'],
-      );
-    });
-
-    test('keep-none returns empty', () {
-      final game = gameWithMinors();
-      final snapshot = _snapshotWithAtWar(['minorA', 'minorB']);
-      expect(
-        minorAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => false,
-        ),
-        isEmpty,
-      );
-    });
-
-    test('returns empty when no at-war minor is present', () {
-      final game = gameWithMinors();
-      final snapshot = _snapshotWithAtWar([_gp1, _tribe1]);
-      expect(
-        minorAtWarPeaceTargetsWhere(game: game, snapshot: snapshot),
-        isEmpty,
-      );
-    });
-
-    test('result is sorted ascending regardless of atWarWith order', () {
-      final game = gameWithMinors();
-      expect(
-        minorAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: _snapshotWithAtWar(['minorC', 'minorA', 'minorB']),
-        ),
-        ['minorA', 'minorB', 'minorC'],
-      );
-    });
-  });
-
-  group('tribeAtWarPeaceTargetsWhere (Refs #3717)', () {
-    Game gameWithTribes() => Game(
-      id: 'g-3717-tribe-peace',
-      worldState: WorldState(
-        turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-        oldWorld: const RegionData(provinces: []),
-        newWorld: const RegionData(provinces: []),
-      ),
-      players: const [Player(id: _gp1, displayName: 'GP1', isHuman: false)],
-      minorNations: const [MinorNation(id: 'minorA', displayName: 'MinorA')],
-      tribes: const [
-        Tribe(id: 'tribeA', displayName: 'TribeA'),
-        Tribe(id: 'tribeB', displayName: 'TribeB'),
-        Tribe(id: 'tribeC', displayName: 'TribeC'),
-      ],
-    );
-
-    test('keep == null keeps every at-war tribe, sorted ascending', () {
-      final game = gameWithTribes();
-      final snapshot = _snapshotWithAtWar([
-        'tribeC',
-        _gp1,
-        'tribeA',
-        'minorA',
-        'tribeB',
-      ]);
-      expect(tribeAtWarPeaceTargetsWhere(game: game, snapshot: snapshot), [
-        'tribeA',
-        'tribeB',
-        'tribeC',
-      ]);
-    });
-
-    test('keeps only tribes matching the predicate, sorted', () {
-      final game = gameWithTribes();
-      final snapshot = _snapshotWithAtWar(['tribeC', 'tribeA', 'tribeB']);
-      expect(
-        tribeAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (factionId) => factionId != 'tribeB',
-        ),
-        ['tribeA', 'tribeC'],
-      );
-    });
-
-    test('never offers a GP or minor even with a keep-all predicate', () {
-      final game = gameWithTribes();
-      final snapshot = _snapshotWithAtWar([_gp1, 'minorA', 'tribeA']);
-      expect(
-        tribeAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => true,
-        ),
-        ['tribeA'],
-      );
-    });
-
-    test('keep-none returns empty', () {
-      final game = gameWithTribes();
-      final snapshot = _snapshotWithAtWar(['tribeA', 'tribeB']);
-      expect(
-        tribeAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => false,
-        ),
-        isEmpty,
-      );
-    });
-
-    test('returns empty when no at-war tribe is present', () {
-      final game = gameWithTribes();
-      final snapshot = _snapshotWithAtWar([_gp1, 'minorA']);
-      expect(
-        tribeAtWarPeaceTargetsWhere(game: game, snapshot: snapshot),
-        isEmpty,
-      );
-    });
-
-    test('result is sorted ascending regardless of atWarWith order', () {
-      final game = gameWithTribes();
-      expect(
-        tribeAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: _snapshotWithAtWar(['tribeC', 'tribeA', 'tribeB']),
-        ),
-        ['tribeA', 'tribeB', 'tribeC'],
-      );
-    });
-  });
-
-  group('nonGreatPowerAtWarPeaceTargetsWhere (Refs #3749)', () {
-    Game gameWithMixedFactions() => Game(
-      id: 'g-3749-non-gp-peace',
-      worldState: WorldState(
-        turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-        oldWorld: const RegionData(provinces: []),
-        newWorld: const RegionData(provinces: []),
-      ),
-      players: const [
-        Player(id: _gp1, displayName: 'GP1', isHuman: false),
-        Player(id: _gp2, displayName: 'GP2', isHuman: false),
-      ],
-      minorNations: const [
-        MinorNation(id: 'minorA', displayName: 'MinorA'),
-        MinorNation(id: 'minorB', displayName: 'MinorB'),
-      ],
-      tribes: const [Tribe(id: _tribe1, displayName: 'Tribe1')],
-    );
-
-    test('keep == null keeps every at-war non-GP faction, sorted ascending', () {
-      final game = gameWithMixedFactions();
-      final snapshot = _snapshotWithAtWar([
-        _tribe1,
-        _gp2,
-        'minorB',
-        _gp1,
-        'minorA',
-      ]);
-      expect(
-        nonGreatPowerAtWarPeaceTargetsWhere(game: game, snapshot: snapshot),
-        ['minorA', 'minorB', _tribe1],
-      );
-    });
-
-    test('keeps an at-war id that is no longer a registered minor or tribe', () {
-      // Pins the `playerById == null` semantics (non-GP), distinct from the
-      // minor/tribe membership collectors: an absorbed faction id still in
-      // `atWarWith` is non-GP and must be retained.
-      final game = gameWithMixedFactions();
-      final snapshot = _snapshotWithAtWar(['absorbedX', _gp1, 'minorA']);
-      expect(
-        nonGreatPowerAtWarPeaceTargetsWhere(game: game, snapshot: snapshot),
-        ['absorbedX', 'minorA'],
-      );
-    });
-
-    test('never offers a Great Power even with a keep-all predicate', () {
-      final game = gameWithMixedFactions();
-      final snapshot = _snapshotWithAtWar([_gp1, _gp2, 'minorA', _tribe1]);
-      expect(
-        nonGreatPowerAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => true,
-        ),
-        ['minorA', _tribe1],
-      );
-    });
-
-    test('keeps only non-GP factions matching the predicate, sorted', () {
-      final game = gameWithMixedFactions();
-      final snapshot = _snapshotWithAtWar(['minorB', _tribe1, 'minorA']);
-      expect(
-        nonGreatPowerAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (factionId) => factionId != 'minorB',
-        ),
-        ['minorA', _tribe1],
-      );
-    });
-
-    test('keep-none returns empty', () {
-      final game = gameWithMixedFactions();
-      final snapshot = _snapshotWithAtWar(['minorA', _tribe1]);
-      expect(
-        nonGreatPowerAtWarPeaceTargetsWhere(
-          game: game,
-          snapshot: snapshot,
-          keep: (_) => false,
-        ),
-        isEmpty,
-      );
-    });
-
-    test('returns empty when only Great Powers are at war', () {
-      final game = gameWithMixedFactions();
-      final snapshot = _snapshotWithAtWar([_gp1, _gp2]);
-      expect(
-        nonGreatPowerAtWarPeaceTargetsWhere(game: game, snapshot: snapshot),
-        isEmpty,
-      );
-    });
-  });
-
-  group('peaceTargetsExcludingBlocker (Refs #3717)', () {
-    test('excludes the blocker and sorts ascending', () {
-      expect(
-        peaceTargetsExcludingBlocker(
-          factionIds: [_gp3, _gp1, _gp2],
-          blocker: _gp2,
-        ),
-        [_gp1, _gp3],
-      );
-    });
-
-    test('null blocker keeps every faction, sorted ascending', () {
-      expect(
-        peaceTargetsExcludingBlocker(
-          factionIds: [_gp3, _gp1, _gp2],
-          blocker: null,
-        ),
-        [_gp1, _gp2, _gp3],
-      );
-    });
-
-    test('blocker absent from the list keeps every faction, sorted', () {
-      expect(
-        peaceTargetsExcludingBlocker(
-          factionIds: [_gp3, _gp1],
-          blocker: _gp2,
-        ),
-        [_gp1, _gp3],
-      );
-    });
-
-    test('empty input returns empty', () {
-      expect(
-        peaceTargetsExcludingBlocker(factionIds: const [], blocker: _gp1),
-        isEmpty,
-      );
-    });
-
-    test('result order is independent of input order', () {
-      expect(
-        peaceTargetsExcludingBlocker(
-          factionIds: [_gp1, _gp3, _gp2],
-          blocker: _gp2,
-        ),
-        peaceTargetsExcludingBlocker(
-          factionIds: [_gp3, _gp2, _gp1],
-          blocker: _gp2,
-        ),
-      );
-    });
-
-    test('accepts a Set (threats.atWarWith) input and excludes blocker', () {
-      expect(
-        peaceTargetsExcludingBlocker(
-          factionIds: <String>{_gp2, _minor1, _gp1},
-          blocker: _gp2,
-        ),
-        [_gp1, _minor1],
-      );
-    });
-  });
-
-  group('scaleWeightedBonus', () {
-    test('returns 0 when weight <= 0.0', () {
-      expect(scaleWeightedBonus(0.0, 45), 0);
-      expect(scaleWeightedBonus(-0.5, 45), 0);
-    });
-
-    test('returns baseConstant exactly when weight == 1.0', () {
-      expect(scaleWeightedBonus(1.0, 45), 45);
-      expect(scaleWeightedBonus(1.0, 360), 360);
-    });
-
-    test('clamps weight > 1.0 to 1.0', () {
-      expect(scaleWeightedBonus(1.5, 45), 45);
-      expect(scaleWeightedBonus(2.0, 100), 100);
-    });
-
-    test('rounds intermediate weights', () {
-      expect(scaleWeightedBonus(0.05, 45), 2);
-      expect(scaleWeightedBonus(0.6, 50), 30);
-    });
-  });
-
-  group('resolvePhaseColonialPressureActive', () {
-    test('true only under COLONIAL', () {
-      expect(
-        resolvePhaseColonialPressureActive(ObserverGoalPhase.colonial),
-        isTrue,
-      );
-      expect(
-        resolvePhaseColonialPressureActive(ObserverGoalPhase.expand),
-        isFalse,
-      );
-      expect(
-        resolvePhaseColonialPressureActive(ObserverGoalPhase.colonialLite),
-        isFalse,
-      );
-      expect(
-        resolvePhaseColonialPressureActive(ObserverGoalPhase.develop),
-        isFalse,
-      );
-    });
-  });
-
-  group('resolvePhaseExpandOrColonialLiteActive', () {
-    test('true under EXPAND and COLONIAL-lite only', () {
-      expect(
-        resolvePhaseExpandOrColonialLiteActive(ObserverGoalPhase.expand),
-        isTrue,
-      );
-      expect(
-        resolvePhaseExpandOrColonialLiteActive(ObserverGoalPhase.colonialLite),
-        isTrue,
-      );
-      expect(
-        resolvePhaseExpandOrColonialLiteActive(ObserverGoalPhase.colonial),
-        isFalse,
-      );
-      expect(
-        resolvePhaseExpandOrColonialLiteActive(ObserverGoalPhase.develop),
-        isFalse,
-      );
-    });
-  });
-
-  group('phase-plan weight projections (Refs #3717)', () {
-    PhasePlanOutcome outcome({
-      ObserverGoalPhase phase = ObserverGoalPhase.expand,
-      PhasePriorityWeights weights = _weights,
-    }) => PhasePlanOutcome(phase: phase, priorityWeights: weights);
-
-    test('each projection returns exactly its named priorityWeights slot', () {
-      final o = outcome();
-      expect(
-        resolvePhaseNewWorldAcquisitionWeight(o),
-        equals(_weights.newWorldAcquisition),
-      );
-      expect(
-        resolvePhaseOldWorldConquestWeight(o),
-        equals(_weights.oldWorldConquest),
-      );
-      expect(
-        resolvePhaseOldWorldCivilianWeight(o),
-        equals(_weights.oldWorldCivilian),
-      );
-      expect(
-        resolvePhaseNewWorldCivilianWeight(o),
-        equals(_weights.newWorldCivilian),
-      );
-    });
-
-    test('projections do not read sibling slots (no field confusion)', () {
-      final o = outcome();
-      // Each of the four distinct slot values is returned by exactly one
-      // projection; a swapped accessor would match the wrong value here.
-      expect(
-        resolvePhaseNewWorldAcquisitionWeight(o),
-        isNot(equals(_weights.oldWorldConquest)),
-      );
-      expect(
-        resolvePhaseOldWorldConquestWeight(o),
-        isNot(equals(_weights.newWorldAcquisition)),
-      );
-      expect(
-        resolvePhaseOldWorldCivilianWeight(o),
-        isNot(equals(_weights.newWorldCivilian)),
-      );
-      expect(
-        resolvePhaseNewWorldCivilianWeight(o),
-        isNot(equals(_weights.oldWorldCivilian)),
-      );
-    });
-
-    test('projections are phase-independent and deterministic', () {
-      final results = <double>{
-        for (final phase in ObserverGoalPhase.values)
-          resolvePhaseNewWorldAcquisitionWeight(outcome(phase: phase)),
-      };
-      expect(results, <double>{_weights.newWorldAcquisition});
-
-      final o = outcome();
-      expect(
-        resolvePhaseOldWorldConquestWeight(o),
-        equals(resolvePhaseOldWorldConquestWeight(o)),
-      );
-    });
-
-    test('flipping a single slot changes only that slot projection', () {
-      const bumped = PhasePriorityWeights(
-        oldWorldConquest: 0.99,
-        newWorldAcquisition: 0.22,
-        oldWorldCivilian: 0.33,
-        newWorldCivilian: 0.44,
-      );
-      final base = outcome();
-      final next = outcome(weights: bumped);
-      expect(
-        resolvePhaseOldWorldConquestWeight(next),
-        isNot(equals(resolvePhaseOldWorldConquestWeight(base))),
-      );
-      expect(
-        resolvePhaseNewWorldAcquisitionWeight(next),
-        equals(resolvePhaseNewWorldAcquisitionWeight(base)),
-      );
-      expect(
-        resolvePhaseOldWorldCivilianWeight(next),
-        equals(resolvePhaseOldWorldCivilianWeight(base)),
-      );
-      expect(
-        resolvePhaseNewWorldCivilianWeight(next),
-        equals(resolvePhaseNewWorldCivilianWeight(base)),
-      );
-    });
-  });
-
 }
