+++ packages/colonizethis_ai/test/planning/planning_invadable_owners_test.dart
@@ -1,34 +1,10 @@
-// Unit tests for `planning_helpers.dart` invadable-province ownership helpers
-// (Refs #3717). Split out of `planning_helpers_test.dart` to keep each test
-// file at or below the repo non-comment line limit
-// (`repo.dart_file_non_comment_line_size`).
-//
-// Pins:
-//   - `anyInvadableProvinceOwnedByMinor` — minor-owner presence scan over the
-//     invadable frontier, `.any` short-circuit, absent/other-owner exclusion
-//   - `anyInvadableProvinceOwnedByGreatPower` — GP-owner presence scan, same
-//     short-circuit / exclusion contract
-//   - `factionOwnsInvadableOldWorldProvince` — single-faction invadable-frontier
-//     ownership scan, `.any` short-circuit, absent/other-owner exclusion
-//   - `addInvadableProvinceMinorOwnersNotAtWar` — minor-owner collector skipping
-//     unowned / non-minor / already-at-war provinces, set de-dup, pre-seeded set
-//   - `mutualExhaustedGpStalemateSideQualifies` — per-side mutual-exhausted
-//     below-quota GP stalemate qualification: min-OW floor, below-quota / stalled
-//     bands, treasury / regiment exhaustion ceilings, unknown-faction exclusion
-//   - `clampPhaseWeightUpperUnit` — `weight > 1.0 ? 1.0 : weight` upper-clamp:
-//     caps above-ceiling weights, passes in-range / negative weights through,
-//     and matches the inline ternary it replaces
+// Unit tests for `planning_invadable_owners.dart` (Refs #3941 topic split).
+// Pins invadable frontier ownership scans/collectors and invadable-blocker gate.
 
 import 'package:colonizethis_ai/src/perception/perception_snapshot.dart';
 import 'package:colonizethis_ai/src/planning/planning_helpers.dart';
 import 'package:colonizethis_ai/src/util/faction_query.dart'
     show isMinorFaction;
-import 'package:colonizethis_data/colonizethis_data.dart'
-    show
-        kMutualExhaustedGpRegimentMax,
-        kMutualExhaustedGpStalemateMinOw,
-        kMutualExhaustedGpTreasuryMax,
-        kObserverConquestMinOwProvincesPerGp;
 import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
@@ -37,6 +13,8 @@ const String _gp2 = 'gp2';
 const String _gp3 = 'gp3';
 const String _tribe1 = 'tribe1';
 const String _minor1 = 'minor1';
+
+
 
 Game _gameWithGps() {
   return Game(
@@ -56,6 +34,8 @@ Game _gameWithGps() {
   );
 }
 
+
+
 AIWorldSnapshot _snapshotWithInvadable(List<String> invadableProvinceIds) {
   return AIWorldSnapshot(
     playerId: _gp1,
@@ -64,41 +44,6 @@ AIWorldSnapshot _snapshotWithInvadable(List<String> invadableProvinceIds) {
     conquest: ConquestSummary(invadableProvinceIdsSorted: invadableProvinceIds),
     economy: const EconomySummary(),
     relations: const {},
-  );
-}
-
-const String _gpExhausted = 'gpExhausted';
-
-// Single-Great-Power game whose treasury and standing-regiment totals are
-// configurable so the mutual-exhausted side qualification can be probed at its
-// economic / military exhaustion ceilings. `mutualExhaustedGpStalemateSideQualifies`
-// reads the side's Old World count from its `ow` argument (not the game), so the
-// fixture only has to back `Game.playerById` and `regimentCountForPlayer`.
-Game _gameWithExhaustedGp({int treasury = 0, int regiments = 0}) {
-  return Game(
-    id: 'g-3717-mutual-exhausted-side',
-    worldState: WorldState(
-      turnState: const TurnState(turnNumber: 1, phase: TurnPhase.orders),
-      oldWorld: const RegionData(provinces: []),
-      newWorld: const RegionData(provinces: []),
-      armies: [
-        Army(
-          id: 'army-$_gpExhausted',
-          ownerId: _gpExhausted,
-          regionId: 'ow',
-          stationedProvinceId: 'ow|home',
-          regimentUnitIds: [for (var i = 0; i < regiments; i++) 'reg$i'],
-        ),
-      ],
-    ),
-    players: [
-      Player(
-        id: _gpExhausted,
-        displayName: 'Exhausted GP',
-        isHuman: false,
-        treasury: treasury,
-      ),
-    ],
   );
 }
 
@@ -544,125 +489,4 @@ void main() {
     });
   });
 
-  group('mutualExhaustedGpStalemateSideQualifies (Refs #3717)', () {
-    // 8 OW provinces sits in the late-stalled "8-9 plateau": at/above the
-    // min-OW floor, below the observer quota (10), and inside the stall band
-    // (<= 9), so it satisfies the three OW gates simultaneously.
-    final int qualifyingOw = kMutualExhaustedGpStalemateMinOw;
-
-    test('true at the min-OW / treasury / regiment exhaustion boundary', () {
-      final game = _gameWithExhaustedGp(
-        treasury: kMutualExhaustedGpTreasuryMax,
-        regiments: kMutualExhaustedGpRegimentMax,
-      );
-      expect(
-        mutualExhaustedGpStalemateSideQualifies(
-          game: game,
-          factionId: _gpExhausted,
-          ow: qualifyingOw,
-        ),
-        isTrue,
-      );
-    });
-
-    test('false below the min-OW floor (even while stalled and below quota)', () {
-      final game = _gameWithExhaustedGp();
-      expect(
-        mutualExhaustedGpStalemateSideQualifies(
-          game: game,
-          factionId: _gpExhausted,
-          ow: kMutualExhaustedGpStalemateMinOw - 1,
-        ),
-        isFalse,
-      );
-    });
-
-    test('false at the observer conquest quota (not below quota / not stalled)', () {
-      final game = _gameWithExhaustedGp();
-      expect(
-        mutualExhaustedGpStalemateSideQualifies(
-          game: game,
-          factionId: _gpExhausted,
-          ow: kObserverConquestMinOwProvincesPerGp,
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when treasury exceeds the exhaustion ceiling', () {
-      final game = _gameWithExhaustedGp(
-        treasury: kMutualExhaustedGpTreasuryMax + 1,
-      );
-      expect(
-        mutualExhaustedGpStalemateSideQualifies(
-          game: game,
-          factionId: _gpExhausted,
-          ow: qualifyingOw,
-        ),
-        isFalse,
-      );
-    });
-
-    test('false when standing regiments exceed the exhaustion ceiling', () {
-      final game = _gameWithExhaustedGp(
-        regiments: kMutualExhaustedGpRegimentMax + 1,
-      );
-      expect(
-        mutualExhaustedGpStalemateSideQualifies(
-          game: game,
-          factionId: _gpExhausted,
-          ow: qualifyingOw,
-        ),
-        isFalse,
-      );
-    });
-
-    test('false for a faction id that does not resolve to a Great Power', () {
-      final game = _gameWithExhaustedGp(
-        treasury: kMutualExhaustedGpTreasuryMax,
-        regiments: kMutualExhaustedGpRegimentMax,
-      );
-      expect(
-        mutualExhaustedGpStalemateSideQualifies(
-          game: game,
-          factionId: 'ghost',
-          ow: qualifyingOw,
-        ),
-        isFalse,
-      );
-    });
-  });
-
-  group('clampPhaseWeightUpperUnit (Refs #3717)', () {
-    test('caps weights above the unit ceiling to 1.0', () {
-      expect(clampPhaseWeightUpperUnit(1.5), 1.0);
-      expect(clampPhaseWeightUpperUnit(2.0), 1.0);
-      expect(clampPhaseWeightUpperUnit(100.0), 1.0);
-    });
-
-    test('returns the boundary weight 1.0 unchanged (strict > ceiling)', () {
-      expect(clampPhaseWeightUpperUnit(1.0), 1.0);
-    });
-
-    test('passes through in-range weights below the ceiling unchanged', () {
-      expect(clampPhaseWeightUpperUnit(0.0), 0.0);
-      expect(clampPhaseWeightUpperUnit(0.05), 0.05);
-      expect(clampPhaseWeightUpperUnit(0.6), 0.6);
-      expect(clampPhaseWeightUpperUnit(0.999), 0.999);
-    });
-
-    test(
-      'does not lower-clamp — negative inputs pass through (callers guard '
-      '<= 0.0 themselves)',
-      () {
-        expect(clampPhaseWeightUpperUnit(-0.5), -0.5);
-      },
-    );
-
-    test('matches the inline ternary it replaces for representative weights', () {
-      for (final w in <double>[-1.0, 0.0, 0.05, 0.5, 1.0, 1.0001, 3.0]) {
-        expect(clampPhaseWeightUpperUnit(w), w > 1.0 ? 1.0 : w);
-      }
-    });
-  });
 }
